Time: 0 ms (100.00%) | Memory: 42.5 MB (64.39%) - LeetSync

+++ 3817-maximum-containers-on-a-ship/maximum-containers-on-a-ship.java
@@ -1,6 +1,6 @@
 class Solution {
     public int maxContainers(int n, int w, int maxWeight) {
-        long totalCapacity = (long) n * n;   
+        long totalCapacity = n * n;   
         long allowedByWeight = maxWeight / w; 
         
         return (int) Math.min(totalCapacity, allowedByWeight);
